% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetFlexTable.R
\name{GetFlexTable}
\alias{GetFlexTable}
\title{GetFlexTable}
\usage{
GetFlexTable(...)
}
\arguments{
\item{...}{\code{\link{ConstructFlexTable}} objects}
}
\description{
Creates a \code{\link{FlexTable}} object using \code{\link{ConstructFlexTable}} objects. Multiple objects can be appended provided they have the same number of rows.
}
\details{
This requires ReporteRs to be installed and ReporteRs requires a java JDK to be installed to work correcly.
}
