\name{prettyTables-package}
\alias{prettyTables-package}
\alias{prettyTables}
\docType{package}
\title{
Create pretty tables 
}
\description{
Tabulation of summary statistics and some utilities for pretty reporting.
}
\details{
\tabular{ll}{
Package: \tab prettyTables\cr
Type: \tab Package\cr
Version: \tab 1.01\cr
Date: \tab 2016-03-08\cr
License: \tab GPL-2\cr
}

\bold{List of key functions:}
\itemize{
\item \code{\link{table_values}}
\item \code{\link{table_two_by_two}}
\item \code{\link{roundWZero}}
\item \code{\link{prettifyPValue}}
}
}
\author{
Peter Dutton

Maintainer: Peter Dutton <peter.dutton@ndorms.ox.ac.uk>
}
