% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roundWZero.R
\name{roundWZero}
\alias{roundWZero}
\title{Round numeric values keeping trailing zeros.}
\usage{
roundWZero(x, digits = 0)
}
\arguments{
\item{x}{a numeric vector.}

\item{digits}{integer indicating the number of decimal places}
}
\description{
\code{roundWZero} rounds the values in its first argument to the specified number of decimal places (default 0).
}
\examples{
roundWZero(c(100,10,1,0.1,0.01,0.001,0.0001),digits = 1)
roundWZero(c(100,10,1,0.1,0.01,0.001,0.0001),digits = -1)

}

