% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_percent.R
\name{table_percent}
\alias{table_percent}
\title{table_percent}
\usage{
table_percent(x, y = NULL, direction = "row", total = "both",
  useNA = "ifany", rnd = 1)
}
\arguments{
\item{x}{first variable to tabulate}

\item{y}{(optional) second variable to tabulate}

\item{direction}{"row" or "col" for horizontal vs vertical percentages}

\item{total}{"both", "row" or "col" for totals columns}

\item{useNA}{"none","ifany","always" passed directly to table}

\item{rnd}{round percentages to this many decimal places. This uses \code{\link{roundWZero}} to force zeros to remain in decimals.}
}
\description{
This is a wrapper function for table. It adds percentages to existing counts normally returned by table. This is only written for 1 or 2 variable.
}
