% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_values.R
\name{table_values}
\alias{table_values}
\title{Creates tables for baseline characteristics or similar tables}
\usage{
table_values(data, var, var.names = NULL, var.order = list(), type,
  strata = NULL, strata.names = NULL, strata.count = TRUE,
  overall = TRUE, count = "n", round = 3, messages = TRUE)
}
\arguments{
\item{data}{Dataset to generate table from}

\item{var}{A vector of the columns (in string form) to be used to generate the table from}

\item{var.names}{Optional. An alternative (cleaned) list of variable names to use in the table. These may contain spaces}

\item{var.order}{Optional. A list containing vectors of strings named using the column names in var. This is used to reorder count data items if the original ordering is not acceptable}

\item{type}{A vector of the method to summarise the var}

\item{strata}{A string of the column to used to stratify on}

\item{strata.names}{Otional. A named vector of alternative strata names}

\item{strata.count}{TRUE/FALSE for displaying strata counts at the top of each column}

\item{overall}{TRUE/FALSE for including an overall column}

\item{count}{"n","miss" or "none" providing the counts, missing values or omitting for each column for numeric variables}

\item{round}{A value or vector for the number of significant figures to report the data to}

\item{messages}{TRUE/FALSE report messages for missing row names in var.order}
}
\value{
Returns a data.frame
}
\description{
Takes a dataset and returns a constructed data.frame summarising the variables of interest. This may be subset by strata and display an overall summary. A shiny user interface is provided (\code{\link{shiny_table_values}}). This runs through creating a table and prints r code to the console or a log file to rerun with an rtf wrapper to export directly to word.
}
\details{
Values labeled as "" or NA are treated as missing values and are thus listed as such. Values labeled as NaN or "NaN" are treated as not applicable and thus are ignored.

Available methods and values for \strong{Type}:
\tabular{cc}{ "miqr" \tab median (Q25,Q75) \cr "miqrr" \tab median (Q25,Q75)[min,max] \cr "mrng" \tab median (Q0,Q100) \cr "avsd" \tab mean (sd) \cr "avci" \tab mean (confidence interval) \cr "st" \tab count \cr "str" \tab count/total \cr "stp" \tab count (percent) \cr "strp" \tab count/total (percent)
}
}
\examples{
# Data
data(iris)

# This function does not work with factors so convert to string.
iris$Species=as.character(iris$Species)

# build the table (without strata)
table_values(iris, var=c("Species","Sepal.Length","Sepal.Width","Petal.Length",
"Petal.Width"),var.names=NULL, type=c("str","miqr","miqrr","miqrr","miqrr"), round=1)

# build the table (with strata)
table_values(iris, var=c("Sepal.Length","Sepal.Width","Petal.Length","Petal.Width"),
var.names=c("Sepal Length","Sepal Width","Petal Length","Petal Width"),
strata="Species", type=c("miqr","miqr","miqr","miqr"), round=1)

}
\seealso{
\code{\link{table_two_by_two}}
}

