% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_two_by_two.R
\name{table_two_by_two}
\alias{table_two_by_two}
\title{Creates tables for baseline characteristics or similar tables}
\usage{
table_two_by_two(data, var, var2, var.name = NULL, var.name2 = NULL,
  var.order = NULL, var.order2 = NULL, type, strata = NULL,
  strata.names = NULL, strata.count = TRUE, overall = TRUE, count = "n",
  round = 3, nmax = 100)
}
\arguments{
\item{data}{Dataset to generate table from}

\item{var}{A character string naming the row variable to stratify by}

\item{var2}{A character string naming the variable to summarise over}

\item{var.name}{Optional. A formatted version of var}

\item{var.name2}{Optional. A formatted version of var2}

\item{var.order}{Optional. A vector of the values of var in the order you want to display them}

\item{var.order2}{Optional. A vector of the values of var2 in the order you want to display them. Note that this creates empty rows when there are no values for a given variable.}

\item{type}{A vector of the method to summarise var2}

\item{strata}{A string of the column to used to stratify on}

\item{strata.names}{Otional. A named vector of alternative strata names}

\item{strata.count}{TRUE/FALSE for displaying strata counts at the top of each column}

\item{overall}{TRUE/FALSE for including an overall column}

\item{count}{"n","miss" or "none" providing the counts, missing values or omitting for each column for numeric variables}

\item{round}{A value or vector for the number of significant figures to report the data to}

\item{nmax}{The initial number of rows assigned to the data.frame. Trimming is performed so this only needs to be changed if the table will have more than 100 rows}
}
\value{
Returns a data.frame
}
\description{
Provide summary statistics on two or three variables. Rows consist of two variables to summaries and columns may consist of an optional stratification variable. There is also an option to print an overall column alongside the stratified summaries.
}
\details{
Available methods and values for \strong{Type}:
\tabular{cc}{ "miqr" \tab median (Q25,Q75) \cr "miqrr" \tab median (Q25,Q75)[min,max] \cr "mrng" \tab median (Q0,Q100) \cr "avsd" \tab mean (sd) \cr "avci" \tab mean (confidence interval) \cr "st" \tab count \cr "str" \tab count/total \cr "stp" \tab count (percent) \cr "strp" \tab count/total (percent)
}
}
\examples{
# Data
data(iris)

# This function does not work with factors so convert to string.
iris$Species=as.character(iris$Species)

# generate a variable to stratify by since iris only has one target
# stratification
iris$Petal.split="Short"
iris$Petal.split[iris$Petal.Length>4.3]="Long"

# build the table
table_two_by_two(iris, var="Species",var2="Sepal.Length",
var.name2="Sepal Length", type="miqr",strata="Petal.split", round=1)

}
\seealso{
\code{\link{table_values}}
}

